Bug fix: changes int64_t to long long int in BestRationalApproximatin.h

+++ include/Math/BestRationalApproximation.h
@@ -17,7 +17,8 @@ namespace gbLAB
     class BestRationalApproximation
     {
     public:
-        typedef int64_t LongIntType;
+        //typedef int64_t LongIntType;
+        typedef long long int LongIntType;
 
     private:
         static std::pair<LongIntType,LongIntType> rat_approx(double f, LongIntType md)
